requirements.txt
fastapi
uvicorn
icalendar
python-dotenv
jinja2
pydantic
email-validator
